Spans | Better Points | `drawframe` command
# Summary:

* Spans
* Better Points
* `drawframe`

## Spans

Spans are an idea I've been floating around in my
head for a while. My ideal use-case is that you
can start and stop this program at will, and in
these cases if you plot the x-axis linearly you
get these ugly gaps where the recording has
paused. The aim of spans is too fix this exact
issue.

However I didn't want to couple the data storage
of the raw pings to this graph drawing technique.
As the simplest way to implement this would be to
have the clock stored into the data, which we
would use to infer gaps. Instead we compute it
based on the standard deviation of the input
signal, once we pass some heuristic threshold we
just assume a gap.

## Better Points

This is just the starting point, but the idea is
that we cache the points to memory rather than
draw them directly. Then at the end of the cache
generation we draw the frame from the cache rather
than relying on the overwriting behaviour we
currently use. As once you get to 10k+ pings the
perf starts to tank.

The other nice side benefit of this is that we can
draw different symbols when too many pings overlap
onto a single "pixel".

## `drawframe`

This is another sub-command which will take a
`.pings` file and omit a finished graph frame,
helpful for debugging and useful for inspecting
the files rather than just the `data` sub-command.