Add Github actions (#4)
# Summary

Main goal of this PR is too add github actions to
run the local suite I was already using
automatically.

* `lint` with golangci-lint (version pinned for
  consistent behaviours)
* `go test -race` (minus the network tests)
* `go build ./...` for all architectures I was
  planning on releasing too.

This required a number of changes, first I
switched from `testify` to `gotest.tools` because
of the much better defaults with `time.Time`
comparisons. However this isn't the end of the
time problems, the next issue is that the compact
binary file format that the pings are saved as
doesn't store the timezone in which it was
recorded, this is fine from a data perspective as
I don't really care about sending these files
across timezones and the timestamps matching.
Timezones are a purely UI/UX concern therefore a
new function `In` can be called on data which will
convert all the timestamps in the data to the
parameter. The actual main binaries will probably
never call this since the local timezone is always
preferred but for tests this ensures that the
string UI/UX can be made consistent by specifying
the time zone that the golden test case is based
against.

-----

Network tests can now use the `env` package to
skip being run: see `ping/api_test.go`

-----

Finally some of the actual drawing tests will
print their respective test diffs straight to
console, also controlled by the `env` package.